+++ CHANGELOG.md
@@ -14,6 +14,9 @@ and **Merged pull requests**. Critical items to know are:
 The versions coincide with releases on pip. Only major versions will be released as tags on Github.
 
 ## [0.0.x](https://github.com/singularityhub/singularity-hpc/tree/main) (0.0.x)
+ - fixing bug to config set/get nested fields, default for wrapper scripts true (0.0.57)
+  - change in behavior, instead of `key:value` for set/add `key value` now also works.
+ - adding support for container.yaml overrides (0.0.56)
  - `SINGULARITY_CONTAINER` and `PODMAN_CONTAINER` support (0.0.55)
  - support for views create, delete, edit, install, uninstall (0.0.54)
   - also including cleanup of module in views on uninstall